+++ admin/subcategory.php
@@ -11,16 +11,30 @@ if(isset($_POST['submit']))
 {
 	$category=$_POST['category'];
 	$subcat=$_POST['subcategory'];
-$sql=mysqli_query($con,"insert into subcategory(categoryid,subcategory) values('$category','$subcat')");
-$_SESSION['msg']="SubCategory Created !!";
+	$result = mysqli_query($con,"SELECT id from subcategory where subcategory = '".$subcat."'");
+	$row_cnt = mysqli_num_rows($result);
+	if($row_cnt > 0)
+	{
+		$_SESSION['msg']="<strong>Sorry!</strong> SubCategory has been already in the list !!";
+	} else {
+		mysqli_query($con,"insert into subcategory(categoryid,subcategory) values('$category','$subcat')");
+		$_SESSION['msg']="<strong>Well done!</strong> SubCategory has been Created !!";
+	}
 
 }
 
 if(isset($_GET['del']))
-		  {
-		          mysqli_query($con,"delete from subcategory where id = '".$_GET['id']."'");
-                  $_SESSION['delmsg']="SubCategory deleted !!";
-		  }
+{
+	$result1 = mysqli_query($con,"SELECT id from products where subCategory = '".$_GET['id']."'");
+	$row_cnt1 = mysqli_num_rows($result1);
+	if($row_cnt1 > 0)
+	{
+		$_SESSION['delmsg']="Could not delete since reference data exist !!";
+	} else {
+		mysqli_query($con,"delete from subcategory where id = '".$_GET['id']."'");
+		$_SESSION['delmsg']="SubCategory deleted !!";
+	}
+}
 
 ?>
 <!DOCTYPE html>
@@ -55,7 +69,7 @@ if(isset($_GET['del']))
 {?>
 									<div class="alert alert-success">
 										<button type="button" class="close" data-dismiss="alert">×</button>
-									<strong>Well done!</strong>	<?php echo htmlentities($_SESSION['msg']);?><?php echo htmlentities($_SESSION['msg']="");?>
+									<?php echo $_SESSION['msg']; ?><?php echo htmlentities($_SESSION['msg']="");?>
 									</div>
 <?php } ?>
